# WIP/RFC: Testing optimizations to broadcast type of calls