Just trying out possibilities of optimizing calls which execute the same thunk on all workers.

With the regular master branch, and 8 workers, sending 10^7 Floats to all workers in parallel takes around 0.72 seconds:

```
function foo(n)
    t=time()
    a=fill(t, 10^n)
    @time @sync begin
        for p in workers()
            @spawnat p global bar=a
        end
    end
    
    #validate that the array has been sent over the network
    for p in workers()
        assert(remotecall_fetch(p, ()->length(bar)) == 10^n) 
        assert(remotecall_fetch(p, ()->bar[1]) == t) 
    end
end
```
```
julia> foo(7)
elapsed time: 0.722923702 seconds (1073779440 bytes allocated, 23.18% gc time)
```


With this PR, which implements a new macro `@spawnall`, the time is around 0.21 seconds
```
function foo(n)
    t=time()
    a=fill(t, 10^n)
    @time @sync @spawnall global bar=a
    
    #validate that the array has been sent over the network
    for p in workers()
        assert(remotecall_fetch(p, ()->length(bar)) == 10^n) 
        assert(remotecall_fetch(p, ()->bar[1]) == t) 
    end
end
```
```
julia> foo(7)
elapsed time: 0.218541978 seconds (134273124 bytes allocated, 7.87% gc time)
```

`@spawnall` serializes the request once and then sends the same buffer to all the workers.