@amitmurthy 2015-02-19
This has become superfluos with the recent change of directly writing large arrays to the socket. Closing it.